requirements.txt
pyttsx3==2.90
speechrecognition==3.8.1
pvporcupine==1.10.0
pyaudio==0.2.11
webbrowser==0.0.1
pystray==0.17.1
Pillow==9.0.1
python-dotenv==0.21.0
